+++ lib/system/ssh/constants.go
@@ -1,4 +1,4 @@
-package system
+package ssh
 
 const LOCALHOST = "127.0.0.1"
 const SSH_PORT = 22
